requirements.txt
discord
python-dotenv
sympy
TenGiphPy
arrow
tzlocal
discord-pretty-help
hololive==0.0.18
pygicord
PyNaCl